Add logrotator spec to several jobs to limit space
Set artifactDaysToKeep and daysToKeep to different settings for
different jobs depending on the context. Tried to overlap lengths with
snaps, so that we can compare artifacts from previous snap results for a
few days before deleting.

We may need to limit by number as well, but this gets us started in
restricting build counts and storage utilization.

I'm also in process of increasing the size of the PVC for satqe-jenkins
and satqe-stage-jenkins